BookStorage-getBooksByAuthor
